backend/requirements.txt
fastapi
uvicorn[standard]
pytest
ccxt
pandas
numpy
supabase
python-dotenv
python-jose[cryptography]
websockets
pydantic
cryptography
redis
slowapi
flask
requests
aiofiles
psutil
TA-Lib
scikit-learn
matplotlib
seaborn